Microservicios/Microservicios/configuracion/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
marshmallow==3.21.1
marshmallow-sqlalchemy==0.29.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
Flask-Cors==4.0.0
Flask-RESTful==0.3.9
Flask-JWT-Extended==4.4.4
requests==2.31.0
routeros-api==0.8.0
paramiko==3.4.0
flask-marshmallow==0.14.0